# nsc-practicum-winter2018-api/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "API for T-Mobile/UST Onboarding tracker",
    "version" : "1.0.5",
    "title" : "Onboarding Tool",
    "contact" : {
      "email" : "jason.panzer@seattlecolleges.edu"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths" : {
...
    "/workflow" : {
      "post" : {
        "summary" : "Creates a new workflow configuration",
        "parameters" : [ {
          "in" : "body",
          "name" : "workflow",
          "description" : "Adds a single workflow.",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Workflow"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Success! Workflow has been created."
          },
          "400" : {
            "description" : "Bad request."
          },
          "401" : {
            "description" : "Authorization information is missing or invalid."
          },
          "403" : {
            "description" : "Operation not authorized."
          },
          "500" : {
            "description" : "Internal server error."
          },
          "501" : {
            "description" : "Service not yet implemented."
          }
        }
      }
    },
    "/workflow/{id}" : {
      "get" : {
        "summary" : "Retrieves a workflow by id.",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "description" : "The id of workflow to get.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success! Workflow has been found.",
            "schema" : {
              "$ref" : "#/definitions/Workflow"
            }
          },
          "400" : {
            "description" : "Bad request."
          },
          "401" : {
            "description" : "Authorization information is missing or invalid."
          },
          "403" : {
            "description" : "Operation not authorized."
          },
          "500" : {
            "description" : "Internal server error."
          },
          "501" : {
            "description" : "Service not yet implemented."
          }
        }
      },
      "put" : {
        "summary" : "Updates an existing workflow.",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "description" : "The id of workflow to amend.",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "workflow",
          "description" : "workflow object",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Workflow"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Success! Workflow has been updated."
          },
          "400" : {
            "description" : "Bad request."
          },
          "401" : {
            "description" : "Authorization information is missing or invalid."
          },
          "403" : {
            "description" : "Operation not authorized."
          },
          "500" : {
            "description" : "Internal server error."
          },
          "501" : {
            "description" : "Service not yet implemented."
          }
        }
      },
      "delete" : {
        "summary" : "Deletes an existing workflow.",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "description" : "The id of workflow to delete.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success! Workflow has been deleted."
          },
          "400" : {
            "description" : "Bad request."
          },
          "401" : {
            "description" : "Authorization information is missing or invalid."
          },
          "403" : {
            "description" : "Operation not authorized."
          },
          "500" : {
            "description" : "Internal server error."
          },
          "501" : {
            "description" : "Service not yet implemented."
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Task" : {
      "type" : "object",
      "description" : "A configuration document for workflow tasks. Workflows are comprised of Task objects, which are stored in a Workflow object's \"tasks\" map.",
      "required" : [ "name" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "example" : "ECCD3A6ED4C54D2DA28C9CDD28F6417E",
          "description" : "Uniquely identifies a task."
        },
        "name" : {
          "type" : "string",
          "example" : "Setup Slack",
          "description" : "The display name of the task."
        },
        "descriptions" : {
          "description" : "A task's instruction set.",
          "type" : "array",
          "items" : {
            "type" : "object",
            "properties" : {
              "viewer" : {
                "type" : "string"
              },
              "instructions" : {
                "type" : "string"
              }
            },
            "example" : {
              "manager" : "Do something to complete this task.",
              "employee" : "Do something to complete this task."
            }
          }
        },
        "viewers" : {
          "type" : "string",
          "enum" : [ "hr", "manager", "employee" ],
          "example" : [ "manager", "employee" ],
          "description" : "Indicates the task viewer based on User type"
        }
      }
    },
    "Workflow" : {
      "type" : "object",
      "description" : "A configuration document for a UTC on-boarding workflow. A workflow document defines all the steps a new hire must complete to be considered officially \"on-boarded.\" An on-boarding blueprint if you will.",
      "properties" : {
        "id" : {
          "type" : "string",
          "example" : "ECCD3A6ED4C54D2DA28C9CDD28F6417E",
          "description" : "Uniquely identifies a workflow."
        },
        "name" : {
          "type" : "string",
          "example" : "Cloud",
          "description" : "The display name of the workflow."
        },
        "description" : {
          "type" : "string",
          "example" : "Onboard a Cloud Developer.",
          "description" : "Describes a workflow's purpose."
        },
        "tasks" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Task"
          },
          "description" : "Associates tasks with a workflow and maintains the order in which Tasks should be completed."
        }
      }
    }
  },
  "host" : "virtserver.swaggerhub.com",
  "schemes" : [ "https" ],
  "basePath" : "/kari_bullard/Cloud-Practicum/1.0.5"
}